Extracted feedback sending to own fragment
