+++ app/src/main/res/layout/fragment_about.xml
@@ -10,21 +10,6 @@
 
 
     <TextView
-        android:id="@+id/txtFeedback"
-        style="@style/TextBody"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Du hast einene Fehler gefunden oder eine Anregung? Schick mir einfach eine Email!" />
-
-    <Button
-        android:id="@+id/btnFeedback"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/txtFeedback"
-        android:layout_marginTop="@dimen/elementMarginSmall"
-        android:text="Feedback geben" />
-
-    <TextView
         android:id="@+id/txtAbout"
         style="@style/TextBody"
         android:layout_width="wrap_content"
